generator/requirements.txt
async-generator==1.10
async-timeout==3.0.1
backports.functools-lru-cache==1.6.1
blis==0.4.1
boto==2.49.0
brotlipy==0.7.0
bz2file==0.98
catalogue==1.0.0
certifi==2020.12.5
click==7.1.2
cymem==2.0.5
cytoolz==0.9.0.1
datasets==1.6.0
decorator==4.4.2
dill==0.2.9
fsspec==0.9.0
html5lib==1.1
huggingface-hub==0.0.2
iniconfig==1.1.1
ipython-genutils==0.2.0
langdetect==1.0.7
mkl-fft==1.3.0
mkl-random==1.2.0
mkl-service==2.3.0
msgpack==0.6.2
msgpack-numpy==0.4.7.1
multiprocess==0.70.11.1
murmurhash==0.28.0
pandas==1.2.3
pandocfilters==1.4.2
plac==0.9.6
pluggy==0.13.1
preshed==3.0.5
protobuf==3.15.6
py==1.10.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pydantic==1.7.4
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
Send2Trash==1.5.0
spacy==2.3.0
spacy-langdetect==0.1.2
spacy-legacy==3.0.6
srsly==1.0.5
testpath==0.4.4
thinc==7.4.1
toml==0.10.2
toolz==0.11.1
torch==1.8.1
tqdm==4.49.0
transformers==4.4.2
ujson==4.0.2
webencodings==0.5.1
wrapt==1.10.11
xxhash==2.0.0
